+++ engine/src/Rendering/LightManager.cpp
@@ -46,7 +46,7 @@ void LightManager::Reallocate(unsigned int required)
 	if (required <= data.allocated)
 		return;
 
-	required = Math::UpperPowerOfTwo(required);
+	required = static_cast<unsigned int>(Math::UpperPowerOfTwo(required));
 
 	// Reserve same amount in entity map
 	entityMap.Reserve(required);
